push

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: akouiss <akouiss@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/17 10:20:48 by akouiss           #+#    #+#             */
-/*   Updated: 2025/10/27 16:45:10 by akouiss          ###   ########.fr       */
+/*   Updated: 2025/10/31 05:04:39 by akouiss          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@ int	ft_atoi(const char *nptr)
 {
 	unsigned int	i;
 	int				sign;
-	int				nbr;
+	long			nbr;
 
 	i = 0;
 	sign = 1;
@@ -32,7 +32,7 @@ int	ft_atoi(const char *nptr)
 		nbr = nbr * 10 + (nptr[i] - '0');
 		i++;
 	}
-	return (nbr * sign);
+	return ((int)(nbr * sign));
 }
 
 // #include <stdio.h>
@@ -60,5 +60,9 @@ int	ft_atoi(const char *nptr)
 // 	printf("xxxxxxxxxxxxxxxxxxxxxx\n");
 // 	printf("ft_atoi = %d\n", ft_atoi("   \n -1"));
 // 	printf("atoi = %d\n", atoi("   \n -1"));
+// 	printf("xxxxxxxxxxxxxxxxxxxxxx\n");
+// 	printf("ft_atoi = %d\n", ft_atoi("   \n -2147483648"));
+// 	printf("xxxxxxxxxxxxxxxxxxxxxx\n");
+// 	printf("atoi = %d\n", atoi("   \n -2147483648"));
 // 	printf("xxxxxxxxxxxxxxxxxxxxxx\n");
 // }
